requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
